preparing for generating classes
disabling older tests for now
started to rename prefix
analysis of common predicates

common predicates are properties
properties are rdf:type rdfs:Class
rdfs:Class is a owl:Class which is in turn a rdfs:Class

Next step will be to create a python class for each rdf:class
used, including properties. When you create an instance of a property
you will create an instance of that specific type. This will create a
new way in rdflib to represent the basic owl/rdf data.

We want to do this because it is similar to the gcc type system and we
need to be prepared to declare instances of classes that are in turn
classes themselves, metaclasses.

A gcc:type_decl defines a specific type of a set of gcc instance
variables.
The type_decl will be a type of node obj. The specific type will again
be a python class. This python class will contain metadata saying it
is from a gcc_typedecl. Instances of that class will be constrained by
the type information extracted from the compiler. Ideally we will be
able to define owl rules to implement some of those type information
constraints or we will need more rules to check the instances.